90% планировщика готово

+++ Hobby/Hobby/DataBase/DbPurposesItem.cs
@@ -1,5 +1,4 @@
 ﻿using SQLite;
-using System;
 
 namespace Hobby.DataBase
 {
@@ -8,16 +7,11 @@ namespace Hobby.DataBase
         [PrimaryKey, AutoIncrement]
         public int ID { get; set; }
 
-        // Внешний ключ — ID задачи из DbItem
         public int TaskID { get; set; }
 
-        // Сколько времени (в мс) планируется потратить
-        public int PlannedTime { get; set; }
-
-        // Крайний срок (DateTime в виде строки)
+        // Теперь long
+        public long PlannedTime { get; set; }
         public string Deadline { get; set; }
-
-        // Снимок TotalWorkTime в момент добавления в план
-        public int InitialTotalTime { get; set; }
+        public long InitialTotalTime { get; set; }
     }
 }
